Split role and assignment items to own mods
This change splits the api::rbac into two submodules with the role and
assignment related structs and functions in their associated modules.
Common items, such as paging-related structs remain in the top-level
rbac module.
