+++ cli/src/action/api/rbac/assignments.rs
@@ -1,0 +1,38 @@
+// Copyright 2018-2020 Cargill Incorporated
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use serde::Deserialize;
+
+use super::Pageable;
+
+#[derive(Deserialize)]
+#[serde(tag = "identity_type", content = "identity")]
+#[serde(rename_all = "lowercase")]
+pub enum Identity {
+    Key(String),
+    User(String),
+}
+
+#[derive(Deserialize)]
+pub struct Assignment {
+    #[serde(flatten)]
+    pub identity: Identity,
+    pub roles: Vec<String>,
+}
+
+impl Pageable for Assignment {
+    fn label() -> &'static str {
+        "assignment list"
+    }
+}
